make /admin/categories requests, local tests and postman tests, fix users logic

+++ ewm-service/src/main/java/ru/practicum/admin_api/users/UserServiceImpl.java
@@ -2,6 +2,7 @@ package ru.practicum.admin_api.users;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.admin_api.users.mapper.UserMapper;
@@ -9,6 +10,7 @@ import ru.practicum.admin_api.users.model.NewUserRequest;
 import ru.practicum.admin_api.users.model.User;
 import ru.practicum.admin_api.users.validation.UserValidator;
 import ru.practicum.dtos.users.UserDto;
+import ru.practicum.exception.exceptions.ApiError;
 
 import java.util.List;
 
@@ -22,7 +24,19 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @Override
     public UserDto addUser(NewUserRequest newUserRequest) {
+
         UserValidator.validateUser(newUserRequest);
+
+        User userWithSameEmail = repository.findByEmail(newUserRequest.getEmail());
+
+        if (userWithSameEmail != null) {
+            throw new ApiError(
+                    HttpStatus.CONFLICT,
+                    "Integrity constraint has been violated.",
+                    "could not execute statement; SQL [n/a]; constraint [uq_email]; nested exception" +
+                            " is org.hibernate.exception.ConstraintViolationException: could not execute statement"
+            );
+        }
 
         User user = new User();
         user.setEmail(newUserRequest.getEmail());
@@ -47,6 +61,13 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @Override
     public void deleteUser(long id) {
+        if (!repository.existsById(id)) {
+            throw new ApiError(
+                    HttpStatus.NOT_FOUND,
+                    "The required object was not found.",
+                    "User with id=" + id + " was not found"
+            );
+        }
         repository.deleteById(id);
     }
 }
